fix: correct values for SMTP protocols (#112)
Previously, "tls" and "ssl" have been valid configuration parameters for
the SMTP protocol. Both have not been valid from the view of Jasper and
Java.
This patch drops "tls" and "ssl" and replaces the valid values with
"smtp" and "smtps".

In addition to that, at least TLS 1.2 is enforced for TLS encrypted
connections.